Warn user about non-existing configured branch
When a user tried to configure Flux against an empty repository or a
branch that did not exist we would report a bogus status and left the
user confused.

This commit makes it possible to configure a branch that should be
checked for existence on creation of a new repository. In case there
is no HEAD found for this branch, it does not exist in git terms, and
we report back a meaningful error message to the user.

As a bonus we also use this configured branch when we create a tag for
our write check, so we do not touch anything the user has not told us
to.