Initialize klog flags so it logs to stderr
This got not caught in CircleCI as it does not perform a rebase
on-top of latest master before testing.

We bumped client-go a while ago which included a switch to klog,
the klog flags need to be intialized before they can be set, as we did
not have this initialization in place, the operator was unable to
start.

As logging to stderr is now the default for klog; only initialize the
flags so it is enabled.