Detect and ignore HelmRelease status update
As the update function receives any update, and we do not want to run
expensive dry-runs for status updates, but do want them for resyncs.

Theoretically this should not result in resyncs being skipped, as a
resync will redeliver a resource that was already known, which
should equal to oldFhr.Status == newFhr.Status.