Log and return early if release is not upgradable
Before this change the operator would follow the installation path if
a Helm release was not in a 'DEPLOYED' state, which would fail later on
as a release with the name already existed.

We now detect what the state of the release is and return (and log)
a descriptive error message in case we can not upgrade the release.