Always list the status of a workload in fluxctl
Before this change the `fluxctl list-workloads` command would only list
the status of a workload if it had containers. But because of how we
integrated HelmReleases into the Flux daemon, a HelmRelease only has
'containers' if we are able to detect images in the .spec.values of the
resource.

After this change it will always list the status of the workload (and
any policies), regardless of the amount of containers it (does not)
have.