Factor out sync bookkeeping
This commit factors out the bookkeeping (maintaining a high water
mark) for sync, and provides two implementations:

 - push a git tag to the upstream (the status quo)
 - make an annotation on the Kubernetes secret used for thhe deploy
   key

The latter is useful if you want to maintain the high water mark
without needing to write to the upstream repo.

Aside from the state implementations, the main body of changes is
daemon (loop) code that refers to the sync tag with code that refers
to the sync.State abstraction.