Elide PolicyTranslator by dealing only in policies
PolicyTranslator exists because the `.flux.yaml` files give commands
for updating annotations, rather than updating policies, following on
from how KubeYAML was used. But the fact that policies appear as
annotations in Kubernetes resources is really up to the code in
cluster/kubernetes.

This commit changes the format of `.flux.yaml` files so a command for
updating _policies_ is supplied, rather than for annotations. All the
translation of policies to annotations is then done in
cluster/kubernetes.

There still needs to be a little code to evaluate the qualified policy
updates (i.e., tag_all) against a particular workload, but this can be
done without reference to whether they end up as annotations.