Factor out raw manifest file operations
It's important that the behaviour in the absence of manifest
generation is the same as before. To make this more obviously so, I've
factored out a resourcestore that _just_ deals with raw manifest
files. Instead of having one implementation that operates in two
distinct modes, there are two implementations.

The more complicated implementation can delegate operations to the
simpler one when it's appropriate, so this is a case of moving methods
around, and not duplicating logic.