Determine ownership by antecedent annotation
Due to Helm creating its own descriptions when the user does not supply
one, the previous approach got shot as we would be unable to determine
ownership after manual helm actions.

The new approach is to check if the resources created by the release
are patched with the correct antecedent annotation, which we create
ourselves and have full control over (and survives rollbacks).

To be able to migrate existing releases to a HelmRelease, empty
(missing) annotations are still handled as true / owned by.