Only require image TS when it is used for sorting
As the semver filter does not make use of the image creation timestamp
for filtering and sorting images, but instead has its own logic based
on the semver spec, the block for automated image updates when a
timestamp was missing for an image with a semver filter policy did more
harm than good.

This commit introduces a `RequiresTimestamp` method on the `Pattern`
interface to indicate if the filter needs the timestamps on images to
be present to give a reliable filtered and sorted output. Which is
implemented in `calculateChanges`, so we only bail on missing
timestamps when they are actually required.