Remove weave-cloudy defaults from fluxctl
Previously, fluxctl would presume that you want to connect via Weave
Cloud, if you provide a token. This commit removes the presumption, in
favour of making both `--url` and `--token` necessary to connect via a
service.

This change is not backward-compatible: some installations using
`fluxctl` will have to supply `--url` (or `FLUX_URL`), and/or change
the env entry `WEAVE_CLOUD_TOKEN` to `FLUX_SERVICE_TOKEN`.

I've left a concession to backward-compatibility in -- port-forwarding
will still work with fluxd that's been installed by Weave Cloud. It's
only one of a set of labels tried, and is for the most part invisible
to the user.