Support both new fluxcd and original annotations
The new prefix for annotations is `fluxcd.io`; but, many files will
use `flux.weave.works`, and we don't want to suddenly not understand
the policies in those.

This commit adds support for interpreting and updating policies as
annotations with either prefix. When updating, if an annotation uses
the old prefix, that will be retained; otherwise, the new prefix is
used.