Fixed rollback.timeout type definition in CRD
Trying to a apply a `HelmRelease` with a set `rollback.timeout` would
result in a validation error: `spec.rollback.timeout in body must be of
type int64: "integer"`.

This was due to the fact that the validation scheme definition contained
a mistake, the `type` should be `integer` and the format is `int64`.