Let people supply Tiller hostname/IP to chart
It's possible to tell the Helm operator to use a hostname or IP
address and port, but this isn't exposed in the chart.

This commit adds chart parameters for the Tiller host and IP. In
theory this would let you point at a Tiller listening on localhost --
but since the chart doesn't also include Tiller in the pod, it's not
quite enough.