Move successful chart fetch signal to reconcile
As the chart source getter methods are used in other places, e.g.
to compare the values checksums, and the condition update causes
the `LastUpdateTime` to renew.

Given this `LastUpdateTime` timestamp is used to determine if an
upgrade should be hold back for rolled back releases, it would
corrupt the 'state machine' and give a false positive, causing a
infinite loop of rollback-release.