Add support for HelmRelease v1 resources
This commit adds support for HelmRelease v1 resources with backwards
compatability for v1beta1 resources, it also reintroduces some of
the removed `integrations/` packages, as those resources have been
removed from the Helm operator.

The v1 resource is prioritized over v1beta1 resources, as Flux is
only aware of the resource kind and not the api version. The
prioritization means that if there are two resources with the same
namespace and name, Flux will only see the v1 resource. The fallback
is based on returned errors from the API, this comes at a cost, as
errors during the first try may actually be relevant.

The version of `github.com/fluxcd/helm-operator` has been pinpointed
on `1.0.0-rc1`.