Fix the resolving of a single v1 HelmRelease
Due to a mistake the name of the resource was used as a namespace
argument, as a result several functionalities relying on this method
did no longer work as it was unable to find the given workload (e.g.
automatic image updates and fluxctl workload selector flags). Passing
along the namespace (as intended) resolves all this.