Use `flux` namespace in `deploy/` examples
This is so that the configuration is not effected by e.g. a namespace
context set in the `kubectl` config of the user, as this caused
problems during the creation of the `ClusterRoleBinding` which was set
to the `default` namespace, while the resources themselves would be
created in the user their configured namespace.

It also simplifies the 'Get started with Kustomize' guide, as a user
does no longer need to create the namespace as a prerequisite but is
still able to overwrite it by adding a `namespace: <targetNamespace>`
to their `kustomization.yaml`.