Add an API to trigger a Git sync
As per the discussion in #1128, there are currently no convenient API
methods that could be used as a webhook interface. For the simple need
of letting Flux know that there are changes to be synced, this API
method can be used. It requires no arguments, only triggers a
Repo.Notify() call, same as /api/flux/v11/notify does if the supplied
payload checks out.