Break cyclical dependency of generated template fs
For `fluxctl install`, we embed a set of templates in the fluxctl
binary by generating a fake filesystem as Go code
(pkg/install/generated_templates.gogen.go).

But the code that generates the fake filesystem also _depends_ on the
fake filesystem, since it does double duty as a program for generating
example manifests from those templates.

This leads to a problem: if anything has upset the generation process,
it's not possible to regenerate the files, since the generation tool
won't build.

(A related problem, worked around elsewhere (#2473), is that it's easy
to introduce spurious differences in the generated code -- which,
since it's checked in -- means painful merges. The solution there
still works here.)

In general we don't care about the generated Go code, only about what
it produces. This commit removes the generated file from the
"uncommitted change" detection (`make check-generated`); and uses
`fluxctl install` to generate the example files, removing the need for
the code generating program to depend on the code it's generating.

NB It's still necessary to commit the generated code, since other
projects depend on the install package.