Docs: fix namespace in `kubectl logs` example
Plus some minor formatting fixes.

There is room to change the 'setup' section of the tutorial to a
'prerequisites' section, with links to all installation tutorials,
as this would ease maintainance long term. This is however a
substantial amount of work as it requires some rewritings in other
parts of the tutorial, and thus outside the scope for this quick fix.