Support additional args for `fluxctl install`
During the development of the initial `fluxctl install` command we
decided against hooking the `AdditionalFluxArgs` up to the command.
Time has however made it clear that:

1. There is a fair amount of people who would like to supply more
   arguments than we currently offer, and the manual edit currently
   required is quite cumbersome.
2. In the end-to-end tests that we are heavily expanding at the
   moment we use `fluxctl install` to bootstrap the Flux daemon in
   various configurations. Non-supported arguments are hacked in with
   `sed`, which  is both very hard to read and quite tricky to
   maintain.

This commit simply adds support for supplying additional arguments
using a `--extra-flux-arguments` flag, which accepts a string array
of `--arg=value`s. For improved UX the alphebetical sorting of flags
has been disabled so that `--extra-flux-arguments` is listed at the
bottom in the `fluxctl install --help` output.