e2e: generate SSH secret per test case
The issue with the setup before this commit surfaced while adding a
new test suite for GPG functionalities.

Due to `fluxctl install` creating a boilerplate secret with the same
name as the secret generated by `run.bash`, it was removed during the
teardown of the test suite which tests the command, as this simply
`kubectl delete -f -`s the output of `fluxctl install`.

This caused my own tests to never get past the point of booting a
new git server instance, as the secret it was trying to mount was
no longer present.

This commit adds two helper methods to `lib/install.bash`,
`generate_ssh_secret` and `delete_generated_ssh_secret` which should
be run during the setup and teardown of tests which require an SSH
key (and `known_hosts`) file to be present. Both functions accept
a parameter to control what the secret should be named, which falls
back to `flux-git-deploy`. This same behaviour has been introduced
to `install_git_srv`.