Return an error if no update commands were run
It's evidently easy to end up with no commands in a config file for
updating manifests, either because you accidentally use
`commandUpdated` with a `patchFile` field, or because the instructions
were unclear, and so on.

Instead of silently running nothing, then complaining that no files
have changed, return an error if nothing got run. This way the person
or system attempting the update will at least see the reason nothing
happened.