Run (or run in trap) deferred commands explicitly
The test runner `bats` installs its own trap(s), and in doing so,
appears to interfere with the trap `defer.bash` sets. So, instead of
setting the trap within the sourced file, require the
deferred-command-runner to be invoked explicitly. `run.bash` can set a
trap to do so; the tests must call `run_deferred` (formerly `on_exit`)
in their teardown function.