Clarify configfile paths by name
This gives the various paths clearer names, by appending `Relative` to
those which are supplied or calculated as relative to another.

The method `ConfigRelativeToWorkingDir()` gives the path to the config
file, starting from the working directory. This encodes two pieces of
information: the working directory aka `--git-path`, and where the
config file is in the git repo.

This is also used as the Source of generated manifests (since it will
always end in the configuration file name, it'll still work in
comparisons e.g., in pkg/daemon/sync.go).