Parse image refs in HelmReleases with >2 elements
This commit fixes the parsing of image references with more than
2 elements in `HelmRelease` resources.

The previous version made the assumption that any given image value
given to `image.ParseRef()` would result in a `image.Ref` with only
the image name set. This assumption is however wrong as the parser
will set the first element of the passed string split by `/` to be
the domain in case there are more than 2 elements in total, resulting
in the first element to be overwritten by a new domain value, if
detected at a later stage.

The solution is to construct the full image ref before calling
`image.ParseRef()`, preventing a partial parse of any image element
snippets that may result in a misinterpretation of the image reference.