e2e: wait for gitsrv to be ready
In the end-to-end tests, we already waited for the git server
to be deployed. However, that's not enough since the repo
may have not be initialized before the tests try to access it.

The change adds a check for the git server to be ready to
receive requests on the master branch.