Correctly scope imagePullSecrets by their namespace
PR #2520 broke the scoping of imagePullSecrets when
allowing to get workloads from all namespaces at once.

The problem is that a secret cache (`seenCreds`) was indexed
by the secret name (making it namespace specific).

Processing all the namespaces at once without adjusting the
scoping opened the door to clashes between secrets from
different namespaces which are called the same.