Raising error if arguments are provided to version and install commands
Aims to conform to the same behavior with other commands: `install` and
`version` ones are not expecting to receive arguments, so raising the
error `errorWantedNoArgs`.
In addition, provided some little test coverage and added the required
flags for the `install` command since `--git-email` was missing but
marked as mandatory.

Pretty sure the change is breaking for users using the installation
method via the CLI instead of the Helm Chart: they could have automation
scripts that are pushing arguments and these could break.

Closes #2808