Use camelCase for config fields
Working in sympathy with spf13/viper means using the same names for
flags as for config file fields (e.g., `git-user`); but it is unusual
to have a YAML document with hyphentated names, rather than camelCase.

To be able to do that, we have to use `viper.BindPFlag`, since that
can be given a flag name. And, to do that reliably, we need to make
sure that the field names do exist as mentioned; the machinery in
cmd/fluxd/config.go is largely in aid of that goal.

This commit also

 - restricts the config file to a single format and name:
   `flux-config.yaml`;
 - Moves the registry-poll-interval out of the config file, since it's
   deprecated.