Let people explicitly include images to scan
Currently there's an option to exclude specific images, which has a
handy default of excluding the Kubernetes components' images. But if
you only ever care about your specific images, it's probably going to
be easier to supply a list of images to _include_.

This commit adds the flag `--registry-include-image`, and a tiny
abstraction so that the exclude/include logic can be applied to image
scanning.