Deprecate fluxctl install in favour of base-config
Lately we have had requests to add flags to `fluxctl install` for
various options; this indicates people would like a non-Helm way to
install flux.

However `fluxctl install` isn't really it -- flags are not a good way
to control the complexity of even a modest Kubernetes configuration
like Flux's.

A better thing to do is to give people a way to get a workable base
config which they can then customise -- indeed, kustomize -- to their
hearts' content.

This commit in effect renames `fluxctl install` to `fluxctl
base-config`, to make it clearer that the expected use is to generate
a base config for further modification. `fluxctl install` remains, for
backward-compatibility, but outputs a deprecation notice to stderr.