Compare last-synced and new resources
`getChangedResources()` in `pkg/daemon/sync.go` was not able to determin
what resources are modified when a manifest file has multiple resource
definitions. This commit replaces `getChangedResources()` with new
`compareResources()` function, which determins resource updates and
deletions by direct comparison of last-synced and newly-synced clones.

`ratchet` now keeps synced resources (in addition to revision). This is
for avoiding generating manifests on every sync, because generating
manifests could be quite expensive.