private-git-host-deploying-via-helm
Added better instructions for how to deploy Flux with Helm using a private git server. The key thing here is *not* to create a ConfigMap `flux-ssh-config` manually, but rather to specify the contexts of that ConfigMap in `ssh.known_hosts` in the Helm Chart.

There's a possible WIBNI to allow the HelmChart to reference a pre-existing  `flux-ssh-config` ConfigMap, but I can't get too excited about that.