Supporting git clone over ssh behind http proxy
Flux, with https://github.com/fluxcd/flux/pull/2279 enhancement, supported git:// [git protocol](https://git-scm.com/book/en/v2/Git-on-the-Server-The-Protocols). However, the downside of the Git protocol is the lack of authentication.
The git over ssh protocol works fine with till we aren't behind a corporate proxy. To make Flux we require to have those extra stanzas in the config file in the root's ssh directory in the flux container. Something similar to:
```
Host github.com
    ProxyCommand socat STDIO PROXY:<proxyIP>:%h:%p,proxyport=<proxyport>,proxyauth=<proxyAuth>
    User git
    Hostname ssh.github.com
    Port 443
    IdentityFile /etc/fluxd/ssh/identity
```

This commit adds the config thru the configMap in ssh.yaml which can be overidden with ssh.config section in the values.yaml.