Add a secretDataKey parameter to the Flux Helm chart
Adding a secretDataKey parameter to the Flux Helm chart
enables the user to override the default data key `identity`
and thereby potentially create the secret from a file with
a different name (such as id_rsa) or use a pre-existing
secret that stores the key value using a different data
key.

Expanding the Helm chart's documentation on the subject of
ssh keys to add some clarity and cover this parameter.