Let productions users table reflow better
It's a pain to add to the production users table, as marked _down_,
since it involves reflowing all the subsequent entries.

Instead, divide the entries into three groups with roughly the same
number of entries, and use HTML markup to put them into three columns,
with line breaks instead of rows. This means adding an entry is
easier; if the columns need to be rebalanced, you can just move the
cell divider `</td><td>`.

Signed-off-by: Michael Bridgen <michael@weave.works>