Don't block ECR if outside AWS
We currently exclude ECR images in regions not explicitly allowed,
even if the cluster is running outside AWS. The reasoning for this
was to save inter-region costs, which makes no sense if the cluster
is in no region at all.

This will scan all ECR images if no region is set and no region
is detected.

Signed-off-by: Roman Orzegowski <roman@mbr-targeting.com>