Roll back Kustomize to v3.8.7
According to #3500, there are breaking changes in the patch release of
Kustomize v3.8.8 which were included in the Flux 1.23.0 release that
upgraded from Kustomize v3.8.4 to v3.8.10.

A PATCH or MINOR release in semver is meant to signal that compatibility
is maintained, but it appears the semver wasn't totally honest and a
behavior was changed in a destructive way for some users.

Other Flux users were looking (#3457) for functionality that was added
in v3.8.5, (so I guess that could have been our first clue semver isn't
strictly honored) rather than revert the update entirely, we'd like to
at least upgrade that far.

Signed-off-by: Kingdon Barrett <kingdon@weave.works>