Disable telemetry
We are removing all of the telemetry code as it is not needed, and being
on by default it may be contrary to CNCF policies around telemetry.

* remove check function

With no telemetry, there's no need for these function implementations or
this syscall in this package anymore.

We need to preserve the cadence of checkpoint.go to avoid disturbing any
thing that would constitute a breaking change, but basically all of the
underlying behavior of check can go away, since every Flux v1 version is
out of date every time, we don't need any remote service or lookup to
tell if an update is needed. (It's always needed)

Users should follow the link to upgrade (eg. the migration guide)

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

Update pkg/checkpoint/checkpoint.go

Signed-off-by: Kingdon Barrett <kingdon@weave.works>
Co-authored-by: Stefan Prodan <stefan.prodan@gmail.com>